backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
# Pin Motor/PyMongo to a compatible pair to avoid private API import errors
motor==3.3.2
pymongo==4.6.3
pydantic==2.9.2
python-dotenv==1.0.1

